Deleted repeatead header file. Refactor usb_printf and oled_printf implementation

+++ firmware/src/middlewares/include/usb_io.h
@@ -1,7 +1,7 @@
 #ifndef USB_IO_H
   #define USB_IO_H
 
-uint8_t usb_printf(const char *, ...);
+int usb_printf(const char *, ...);
 uint32_t rtc_get_timestamp();
 void usb_print_memory_info(void);
